+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.1]
+- Added support for custom `width` and `height` dimensions.
+- Added native SVG wobble animations for characters when `difficulty` or `globalNoise` is greater than 3.
+
 ## [0.1.0]
 - Initial release.
 - Added pure TS implementation of 5x7 bitmap rendering.